Move to geoparquet_writer mod and more tweaks
